requirements.txt
requests
wget
Pillow